+++ ingestion/parsers/promoters/custom_script/__init__.py
@@ -1,22 +1,11 @@
-def get_promoters_from_string(promoter_string):
-    return [promoter.strip() for promoter in promoter_string.split(',') if promoter != 'nan']
+import json
 
 
 def custom_script(data):
-    new_dict = {}
-    for split in data:
-        new_dict[split] = {}
-        for neuron in data[split]:
-            for source in data[split][neuron]:
-                for promoter in get_promoters_from_string(data[split][neuron][source]['promoter']):
-                    if promoter in new_dict[split]:
-                        new_dict[split][promoter][source]['cellsByLineaging'].append(neuron)
-                    else:
-                        new_dict[split][promoter] = {
-                            source: {
-                                'uid': promoter,
-                                'cellsByLineaging': [neuron]
-                            }
-                        }
-
-    return new_dict
+    for tp in data:
+        for promoter in data[tp]:
+            for source in data[tp][promoter]:
+                if 'cellsByLineaging' in data[tp][promoter][source]:
+                    data[tp][promoter][source]['cellsByLineaging'] = json.dumps(
+                        data[tp][promoter][source]['cellsByLineaging'])
+    return data
